feat: add 'Add Liquidity' button to pool page
