+++ apps/web/src/app/(evm)/[chainId]/pool/v2/[address]/(landing)/page.tsx
@@ -15,7 +15,7 @@ export default async function PoolPage({
     {
       revalidate: 60 * 15,
     },
-  )()) as NonNullable<V2Pool>
+  )()) as V2Pool
 
   return <PoolPageV2 pool={pool} />
 }
