+++ apps/web/src/app/(evm)/[chainId]/pool/v2/[address]/layout.tsx
@@ -1,9 +1,6 @@
 import { getV2Pool } from '@sushiswap/graph-client/data-api'
-import { Container } from '@sushiswap/ui'
 import { unstable_cache } from 'next/cache'
-import { headers } from 'next/headers'
-import { PoolHeader } from 'src/ui/pool/PoolHeader'
-import { ChainKey, isChainId } from 'sushi'
+import { ChainId } from 'sushi'
 import notFound from '../../../not-found'
 
 export const metadata = {
@@ -17,13 +14,11 @@ export default async function Layout({
   children: React.ReactNode
   params: { chainId: string; address: string }
 }) {
-  const { chainId, address } = params
-  const _chainId = Number(chainId)
-  if (!isChainId(_chainId)) {
-    return notFound()
-  }
+  const { chainId: _chainId, address } = params
+  const chainId = +_chainId as ChainId
+
   const pool = await unstable_cache(
-    async () => getV2Pool({ chainId: _chainId, address }),
+    async () => getV2Pool({ chainId, address }),
     ['pool', `${chainId}:${address}`],
     {
       revalidate: 60 * 15,
@@ -34,27 +29,6 @@ export default async function Layout({
   if (!pool || pool.id === '42161:0x0a4f9962e24893a4a7567e52c1ce37d5482365de') {
     return notFound()
   }
-  const headersList = headers()
-  const referer = headersList.get('referer')
-  return (
-    <>
-      <Container maxWidth="5xl" className="pt-10 px-4">
-        <PoolHeader
-          backUrl={
-            referer?.includes('/pool?')
-              ? referer?.toString()
-              : `/${ChainKey[_chainId]}/explore/pools`
-          }
-          address={pool.address}
-          pool={pool}
-          apy={{ rewards: pool?.incentiveApr, fees: pool?.feeApr1d }}
-        />
-      </Container>
-      <section className="flex flex-col flex-1 mt-4">
-        <div className="bg-gray-50 dark:bg-white/[0.02] border-t border-accent pt-10 pb-20 h-full">
-          {children}
-        </div>
-      </section>
-    </>
-  )
+
+  return <>{children}</>
 }
