+++ apps/web/src/app/(evm)/[chainId]/pool/v2/[address]/migrate/page.tsx
@@ -1,0 +1,40 @@
+import { V2Pool, getV2Pool } from '@sushiswap/graph-client/data-api'
+import { unstable_cache } from 'next/cache'
+import {
+  PoolPositionProvider,
+  PoolPositionRewardsProvider,
+  PoolPositionStakedProvider,
+} from 'src/ui/pool'
+import { ConcentratedLiquidityProvider } from 'src/ui/pool/ConcentratedLiquidityProvider'
+import { MigrateTab } from 'src/ui/pool/MigrateTab'
+import { ChainId } from 'sushi/chain'
+
+export default async function MigrateV2PoolPage({
+  params,
+}: {
+  params: { chainId: string; address: string }
+}) {
+  const { chainId: _chainId, address } = params
+  const chainId = +_chainId as ChainId
+  const pool = (await unstable_cache(
+    async () => getV2Pool({ chainId, address }),
+    ['pool', `${chainId}:${address}`],
+    {
+      revalidate: 60 * 15,
+    },
+  )()) as V2Pool
+
+  return (
+    <div className="flex flex-col gap-6">
+      <PoolPositionProvider pool={pool}>
+        <PoolPositionStakedProvider pool={pool}>
+          <PoolPositionRewardsProvider pool={pool}>
+            <ConcentratedLiquidityProvider>
+              <MigrateTab pool={pool} />
+            </ConcentratedLiquidityProvider>
+          </PoolPositionRewardsProvider>
+        </PoolPositionStakedProvider>
+      </PoolPositionProvider>
+    </div>
+  )
+}
