+++ apps/web/src/app/(evm)/[chainId]/pool/v3/[address]/layout.tsx
@@ -1,9 +1,6 @@
 import { getV3Pool } from '@sushiswap/graph-client/data-api'
-import { Container } from '@sushiswap/ui'
 import { unstable_cache } from 'next/cache'
-import { headers } from 'next/headers'
-import { PoolHeader } from 'src/ui/pool/PoolHeader'
-import { ChainId, ChainKey } from 'sushi/chain'
+import { ChainId } from 'sushi/chain'
 import notFound from '../../../not-found'
 
 export const metadata = {
@@ -31,27 +28,5 @@ export default async function Layout({
     return notFound(chainId)
   }
 
-  const headersList = headers()
-  const referer = headersList.get('referer')
-  return (
-    <>
-      <Container maxWidth="5xl" className="pt-10 px-4">
-        <PoolHeader
-          backUrl={
-            referer?.includes('/pool?')
-              ? referer?.toString()
-              : `/${ChainKey[chainId]}/explore/pools`
-          }
-          address={address}
-          pool={pool}
-          apy={{ rewards: pool.incentiveApr, fees: pool.feeApr1d }}
-        />
-      </Container>
-      <section className="flex flex-col flex-1 mt-4">
-        <div className="bg-gray-50 dark:bg-white/[0.02] border-t border-accent pt-10 pb-20 h-full">
-          {children}
-        </div>
-      </section>
-    </>
-  )
+  return <>{children}</>
 }
