+++ apps/web/src/app/(evm)/[chainId]/pool/v3/[address]/smart/(overview)/layout.tsx
@@ -6,6 +6,7 @@ import {
   LinkInternal,
 } from '@sushiswap/ui'
 import React from 'react'
+import { ChainId, ChainKey } from 'sushi/chain'
 
 export default async function Layout({
   children,
@@ -14,16 +15,18 @@ export default async function Layout({
   children: React.ReactNode
   params: { chainId: string; address: string }
 }) {
+  const { chainId: _chainId, address } = params
+  const chainId = +_chainId as ChainId
   return (
-    <div className="flex flex-col gap-2">
-      <Container maxWidth="5xl" className="px-2 sm:px-4">
+    <div>
+      <Container maxWidth="5xl" className="py-10 px-4 flex flex-col gap-4">
         <LinkInternal
-          href={`/${params.chainId}/pool/v3/${params.address}`}
+          href={`/${ChainKey[chainId]}/pool/v3/${address}`}
           className="text-blue hover:underline text-sm"
         >
           ← Pool details
         </LinkInternal>
-        <CardHeader className="!px-0 !pb-0">
+        <CardHeader className="!p-0">
           <CardTitle>Available Strategies</CardTitle>
           <CardDescription>
             Choose one of the following strategies:
