+++ apps/web/src/ui/pool/SmartPoolsTable.tsx
@@ -322,7 +322,7 @@ const COLUMNS = [
                   onClick={(e) => e.stopPropagation()}
                   shallow={true}
                   className="flex items-center"
-                  href={`/${row.original.chainId}/pool/v3/${row.original.poolAddress}/positions/create`}
+                  href={`/${row.original.chainId}/pool/v3/${row.original.poolAddress}/create`}
                 >
                   <PlusIcon width={16} height={16} className="mr-2" />
                   Create position
@@ -336,7 +336,7 @@ const COLUMNS = [
                         onClick={(e) => e.stopPropagation()}
                         shallow={true}
                         className="flex items-center"
-                        href={`/${row.original.chainId}/pool/v3/${row.original.poolAddress}/smart/${row.original.id}`}
+                        href={`/${row.original.chainId}/pool/v3/${row.original.poolAddress}/smart/${row.original.address}`}
                       >
                         <span className="relative">
                           <LightBulbIcon
